api-ticktick/requirements.txt
fastmcp>=2.3.0
requests>=2.31.0
python-dotenv>=1.0.0
uvicorn>=0.30.0
starlette>=0.37.0
